requirements.txt
boto3>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
pillow>=10.0.0
